requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
Flask==3.0.0
python-dotenv==1.0.0
schedule==1.2.0
APScheduler==3.10.4
psutil==5.9.6

